HOMEB-34 : First Draft

+++ src/main/java/com/epitech/homepedia/dto/DataDTO.java
@@ -5,13 +5,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @RequiredArgsConstructor
 @AllArgsConstructor
 public class DataDTO {
     private String code;
-    private BigDecimal price;
+    private Integer price;
 }
